more asset thread stuff
