+++ include/sys/fs/asset_thread.h
@@ -1,3 +1,6 @@
+#include "PR/os.h"
+#include "PR/ultratypes.h"
+
 enum AssetLoadType {
     ASSET_TYPE_FILE = 0,
     ASSET_TYPE_ALLOCATED_FILE = 1,
@@ -77,6 +80,31 @@ extern OSMesgQueue assetLoadThreadSendQueue, //send load requests to asset threa
     assetLoadThreadRecvQueue; //receive acknowledgement from asset thread
 
 extern u8 D_800AE29D, D_800AE29E;
+
+struct UnkStruct8000ADF0 {
+    s16 unk0;
+};
+extern struct UnkStruct8000ADF0 *D_800ACBC8;
+struct UnkStruct8000ADF0 *func_8000ADF0(s32 *, s32 *, s32, s32);
+
+struct UnkStruct8000B010 {
+    s16 unk0;
+};
+extern struct UnkStruct8000B010 *D_800AE1D0;
+struct UnkStruct8000B010 *func_8000B010(s32 *, s32 *, s32, s32);
+
+extern u8 gDisableObjectStreamingFlag;
+
+extern u8 D_800AE29D, D_800AE29E;
+
+void asset_thread_main(void*);
+
+extern s32 *D_800ACBB8, *D_800ACBD0;
+
+extern s32 *D_800AE1C0, *D_800AE1D8;
+
+extern u64 *assetThreadStackEnd; // end of stack
+extern OSThread assetThread;
 
 void queue_alloc_load_file(void **dest, s32 fileId);
 void queue_load_file_to_ptr(void **dest, s32 fileId);
